start setpoint module
